Third digit increases clicks by the absolute value of 2nd minus 3rd digit.
Unfortunately realized writing this that the click calculation methods are all
wrong. This solution assumes we always take the shortest path to our next number
but of course we may have to take the long way around, depending on our
direction and which digit is greater.

I'll fix it someday, or not.